[Table Visualization] restore pagination to table vis
* add pagination

patically resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2305
