+++ src/plugins/vis_type_table/public/utils/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from './convert_to_formatted_data';
+export * from './use_pagination';
